Missed custom create funcitonality for KeyFile Implemented
