+++ src/keystore/serialize/mod.rs
@@ -6,6 +6,7 @@ mod byte_array;
 mod crypto;
 mod error;
 
+use self::address::try_extract_address;
 use self::crypto::Crypto;
 use self::error::Error;
 use super::{CIPHER_IV_BYTES, Cipher, KDF_SALT_BYTES, Kdf, KeyFile};
@@ -13,10 +14,11 @@ use super::core::{self, Address};
 use super::util;
 use chrono::prelude::UTC;
 use rustc_serialize::{Decodable, Decoder, Encodable, Encoder, json};
-use std::fs::File;
-use std::io::Write;
+use std::fs::{self, File};
+use std::io::{Read, Write};
 use std::path::Path;
 use uuid::Uuid;
+
 
 /// Keystore file current version used for serializing
 pub const CURRENT_VERSION: u8 = 3;
@@ -38,6 +40,39 @@ impl KeyFile {
         let data = json::encode(self)?;
         file.write_all(data.as_ref()).ok();
         Ok(())
+    }
+
+    /// Search of `KeyFile` by specified `Address`
+    ///
+    /// # Arguments
+    ///
+    /// * `path` - path with keystore files
+    ///
+    pub fn search_by_address<P: AsRef<Path>>(addr: &Address, path: P) -> Result<KeyFile, Error> {
+        let entries = fs::read_dir(path)?;
+
+        for entry in entries {
+            let path = entry?.path();
+
+            if path.is_dir() {
+                continue;
+            }
+
+            let mut file = fs::File::open(path)?;
+            let mut content = String::new();
+
+            if file.read_to_string(&mut content).is_err() {
+                continue;
+            }
+            match try_extract_address(&content) {
+                Some(a) if a == *addr => {
+                    return Ok(json::decode::<KeyFile>(&content)?);
+                }
+                _ => continue,
+            }
+        }
+
+        Err(Error::NotFound)
     }
 }
